requirements.txt
django-analytics==0.0.1
django-bootstrap3==6.1.0
django-geckoboard==1.2.8
django-google-analytics==1
Pillow==2.9.0
python-memcached==1.54
six==1.9.0
sorl-thumbnail==12.2
virtualenv==13.0.3
wheel==0.24.0
